log 관련 메소드 logdbbean으로 이동
