+++ src/main/java/handler/SvcFormHandler.java
@@ -174,12 +174,9 @@ public class SvcFormHandler {
 		for(int i=0; i<usertrip.size(); i++) {		
 			TripDataBean tripDto = usertrip.get(i);
 			tripDto.setCoordinate(tripDto.getTrip_id());
-			tripDto.setReview_members(tripDto.getTrip_id());
-			
-			trip.add(tripDto);
-			
-		}	
-		
+			tripDto.setReview_members(tripDto.getTrip_id());			
+			trip.add(tripDto);			
+		}		
 		request.setAttribute("trip", trip);
 		return new ModelAndView("svc/review");		
 	}
